backend/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
blinker==1.7.0
build==1.1.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
coloredlogs==15.0.1
dataclasses-json==0.6.4
Deprecated==1.2.14
dnspython==2.5.0
email-validator==2.1.0.post1
exa-py==1.0.8
exceptiongroup==1.2.0
faiss-cpu==1.7.4
fastapi==0.109.2
filelock==3.13.1
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2024.2.0
gevent==24.2.1
gitdb==4.0.11
GitPython==3.1.42
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.1
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.62.0
grpcio-status==1.62.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.21.3
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib_resources==6.1.2
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kubernetes==29.0.0
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.28
langchain-google-genai==0.0.9
langsmith==0.1.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.0
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.1
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
orjson==3.9.13
overrides==7.7.0
packaging==23.2
pandas==2.2.1
pillow==10.2.0
pinecone-client==3.2.2
posthog==3.4.2
proto-plus==1.23.0
protobuf==4.25.3
pulsar-client==3.4.0
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.1
pydantic-extra-types==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.16.2
pydeck==0.8.1b0
Pygments==2.17.2
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.0.0
python-dateutil==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.9
pyttsx3==2.90
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.13.0
sentence-transformers==2.6.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.27
starlette==0.36.3
streamlit==1.31.1
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.4.0
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.2.2
tornado==6.4
tqdm==4.66.2
transformers==4.39.3
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.27.1
uvloop==0.19.0
validators==0.22.0
watchfiles==0.21.0
websocket==0.2.1
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
zope.event==5.0
zope.interface==6.2
